% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gcomp}
\alias{gcomp}
\title{Calculate gravity component}
\usage{
gcomp(dem, dem.info, dz, sgx, sgy, sgdz, umbrella = F, sghouse, umbrella.max)
}
\arguments{
\item{dem}{DEM in UTM}

\item{dem.info}{information to the used DEM, see details}

\item{dz}{vector of thickness of each layer horizon, number of values equals number of layers used for calculation}

\item{sgx,sgy}{coordinates of gravimeter location}

\item{sgdz}{high of gravimeter sensor above DEM}

\item{umbrella}{logical, wether to compute excluding the umbrella effect, default TRUE}

\item{sghouse}{coordinates of the gravimeter house, used as umbrella surface}

\item{umbrella.max}{maximal depth until which umbrella effect will be included}
}
\description{
Calculates the gravity components (100percent values) for a given DEM using a nested approach
}
\details{
missing
}
\examples{
example MISSING
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}

