% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gdisp}
\alias{gdisp}
\title{Displaying gravity results}
\usage{
gdisp(grav1.array, grav1.name, grav2.array, grav2.name, vertsum = F, layer,
  onlydif = F, cuts = c(-10, -2, -1, -0.5, -0.2, -0.1, -0.01, -0.001, 0,
  0.001, 0.01, 0.1, 0.2, 0.5, 1, 2, 10))
}
\arguments{
\item{vertsum}{logical, whether values will show vertical sums (TRUE) or only one layer (FALSE), default FALSE}

\item{layer}{layer to show, only affective when vertsum=FALSE}

\item{onlydif}{logical, should absolute values be shown of layer x grav1 and grav2 (FALSE) or should each layer be shown as difference grav1 - grav2, default FALSE}

\item{cuts}{determinating breaks categories of input values}

\item{grav1/2.array}{array of values [rows, columns, layers] containing data (usually gravity components or gravity signals), both must have the same dimensions}

\item{grav1/2.name}{name of devices used in graphic}
}
\description{
Visualizes gravity data from two devices in varies posibilities
}
\details{
missing
}
\examples{
example MISSING
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}

