% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hydmodconcep}
\alias{hydmodconcep}
\title{Calculate Hydrological Model}
\usage{
hydmodconcep(data.in, parameters, output = "absolut")
}
\arguments{
\item{data.in}{input data as a data.frame with columns structured: $time, $snowH, $precip, $ETo}

\item{parameters}{set of parameters classifying the different soil layers}

\item{output}{set output to "absolut": absolute values of each storage at each timestep, "dif": values with differences to value at t0, "meanval": values with diffeence to mean value of each storage compartment.}
}
\description{
Various single storage buckets conected in 1D with a conceptional approach
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}

