% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_dem}
\alias{read_dem}
\title{Read Digital Elevation Models}
\usage{
read_dem(dempath, filename)
}
\arguments{
\item{dempath}{Path of input DEM-file.}

\item{filename}{DEM-file in .acs-format (see details)}
}
\description{
Read DEM's and output information files for further usage.
}
\details{
So far only it is only possible to read DEM-file in .asc arquitecture format.
}
\examples{
missing
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}

