% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_hydrus2d}
\alias{read_hydrus2d}
\title{Read Hydrus 2D output data}
\usage{
read_hydrus2d(folder, timesteps, vertical_nodes, horizontal_nodes, t_printout,
  plotting = F, timestamps, px)
}
\arguments{
\item{folder}{foldername of project to read}

\item{timesteps}{number of modeled timesteps}

\item{vertical_nodes}{number of nodes used in the vertical model discretisation}

\item{horizontal_nodes}{number of nodes used in the horizontal model discretisation}

\item{t_printout}{number of model printout times}

\item{plotting}{indicate if standard parameters should be plotted (TRUE); default is FALSE}

\item{timestamps}{vector of timestamps of the modeled timeseries. if not provided output time will be in counts of timesteps}

\item{px}{number of horizontal node, where the vertical profile should be analyzed}
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}

